requirements.txt
streamlit==1.32.0
pandas
numpy
plotly
scikit-learn
xgboost
lightgbm
statsmodels
tensorflow-cpu
joblib
requests
apscheduler